requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
openai==1.12.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-telegram-bot==20.8
redis==5.0.4
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
